Update sqlite3 to version 3.45.2 / rev 150 via SR 1158281
https://build.opensuse.org/request/show/1158281
by user rmax + anag+factory
- Update to release 3.45.2:
  * Added the SQLITE_RESULT_SUBTYPE property for application-
    defined SQL functions.
  * Enhancements to the JSON SQL functions
  * Add the FTS5 tokendata option to the FTS5 virtual table.
  * The SQLITE_DIRECT_OVERFLOW_READ optimization is now enabled by
    default.
  * Query planner improvements
  * Increase the default value for SQLITE_MAX_PAGE_COUNT from
    1073741824 to 4294967294.
  * Enhancements to the CLI
  * Restore the JSON BLOB input bug, and promise to support the
    anomaly in subsequent releases, for backward compatibility.
  * Fix the PRAGMA integrity_check command so that it works on
    read-only databases that contain FTS3 and FTS5 tables.
  * Fix issues associated with processing corrupt JSONB inputs.
  * Fix a long-standing bug in whic

+++ packages/s/sqlite3/sqlite3.spec
@@ -1,7 +1,7 @@
 #
 # spec file for package sqlite3
 #
-# Copyright (c) 2023 SUSE LLC
+# Copyright (c) 2024 SUSE LLC
 #
 # All modifications and additions to the file contributed by third parties
 # remain the property of their copyright owners, unless otherwise agreed
@@ -17,12 +17,12 @@
 
 
 %define oname sqlite
-%define tarversion 3440200
-%define docversion 3440200
+%define tarversion 3450200
+%define docversion 3450200
 %bcond_with icu
 %bcond_without check
 Name:           sqlite3
-Version:        3.44.2
+Version:        3.45.2
 Release:        0
 Summary:        Embeddable SQL Database Engine
 License:        SUSE-Public-Domain
@@ -31,7 +31,7 @@ URL:            https://www.sqlite.org/
 Source0:        https://www.sqlite.org/2023/sqlite-src-%{tarversion}.zip
 Source1:        baselibs.conf
 Source2:        https://www.sqlite.org/2023/sqlite-doc-%{docversion}.zip
-Patch0:         sqlite3-rtree-i686.patch
+Patch0:         sqlite3-float-i586.patch
 BuildRequires:  automake
 BuildRequires:  libtool
 BuildRequires:  pkgconfig
@@ -129,7 +129,15 @@ other documentation found on sqlite.org. The files can be found in
 %{_docdir}/%{name}-doc.
 
 %prep
-%autosetup -p1 -n sqlite-src-%{tarversion} -a2
+# Version and %tarversion need to match, but %docversion might be different,
+IFS=. read a b c d <<< "%version"
+if [ "%tarversion" != $(printf "%1d%02d%02d%02d" $a $b $c $d) ]
+then
+	echo "Version %version does not match tarversion %tarversion."
+	exit 1
+fi
+
+%autosetup -p0 -n sqlite-src-%{tarversion} -a2
 
 rm -v sqlite-doc-%{docversion}/releaselog/current.html
 ln -sv `echo %{docversion} | sed "s/\./_/g"`.html sqlite-doc-%{docversion}/releaselog/current.html
